in the middle of working on the move of diagram
