+++ CSETWebApi/CSETWeb_Api/BusinessLogic/BusinessLogic.csproj
@@ -242,7 +242,8 @@
     <Compile Include="BusinessManagers\AssessmentManager.cs" />
     <Compile Include="BusinessManagers\AssessmentModeManager.cs" />
     <Compile Include="BusinessManagers\ContactsManager.cs" />
-    <Compile Include="BusinessManagers\DiagramManager.cs" />
+    <Compile Include="BusinessManagers\Diagram\DiagramDifferenceManager.cs" />
+    <Compile Include="BusinessManagers\Diagram\DiagramManager.cs" />
     <Compile Include="BusinessManagers\DocumentManager.cs" />
     <Compile Include="BusinessManagers\FrameworkManager.cs" />
     <Compile Include="BusinessManagers\FullAnswer.cs" />
